+++ app/build.gradle.kts
@@ -82,8 +82,7 @@ dependencies {
     implementation(project(":feature:my-info:update-profile"))
     implementation(project(":feature:my-info:update"))
 
-
-    implementation(project(":feature:update-routine:routine"))
     implementation(project(":feature:update-routine:routine-selection"))
+    implementation(project(":feature:update-routine:routine-set"))
 
 }
