+++ app/src/main/java/com/gradation/lift/navigation/graph/UpdateRoutineGraphBuilder.kt
@@ -5,8 +5,8 @@ import androidx.annotation.RequiresApi
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.navigation
-import com.gradation.lift.feature.update_routine.routine.updateRoutineRoutineScreen
 import com.gradation.lift.feature.update_routine.routine_selection.updateRoutineRoutineSelectionScreen
+import com.gradation.lift.feature.update_routine.routine_set.updateRoutineRoutineSetScreen
 import com.gradation.lift.navigation.Router.UPDATE_ROUTINE_GRAPH_NAME
 import com.gradation.lift.navigation.Router.UPDATE_ROUTINE_ROUTINE_SELECTION_ROUTER_NAME
 
@@ -20,7 +20,7 @@ fun updateRoutineGraphBuilder(
         startDestination = UPDATE_ROUTINE_ROUTINE_SELECTION_ROUTER_NAME,
     ) {
         updateRoutineRoutineSelectionScreen(navController, this)
-        updateRoutineRoutineScreen(navController, this)
+        updateRoutineRoutineSetScreen(navController, this)
     }
 }
