+++ core/navigation/src/main/java/com/gradation/lift/navigation/navigation/UpdateRoutineNavigation.kt
@@ -2,8 +2,8 @@ package com.gradation.lift.navigation.navigation
 
 import androidx.navigation.NavController
 import com.gradation.lift.navigation.Router.HOME_GRAPH_NAME
-import com.gradation.lift.navigation.Router.UPDATE_ROUTINE_ROUTINE_ROUTER_NAME
 import com.gradation.lift.navigation.Router.UPDATE_ROUTINE_ROUTINE_SELECTION_ROUTER_NAME
+import com.gradation.lift.navigation.Router.UPDATE_ROUTINE_ROUTINE_SET_ROUTER_NAME
 
 fun NavController.navigateUpdateRoutineGraphToHomeGraph() {
     this.navigate(HOME_GRAPH_NAME) {
@@ -15,13 +15,13 @@ fun NavController.navigateUpdateRoutineGraphToHomeGraph() {
 }
 
 
-fun NavController.navigateRoutineSelectionToRoutine() {
-    this.navigate(UPDATE_ROUTINE_ROUTINE_ROUTER_NAME)
+fun NavController.navigateRoutineSelectionToRoutineSet() {
+    this.navigate(UPDATE_ROUTINE_ROUTINE_SET_ROUTER_NAME)
 }
 
-fun NavController.navigateRoutineToRoutineSelection() {
+fun NavController.navigateRoutineSetToRoutineSelection() {
     this.navigate(UPDATE_ROUTINE_ROUTINE_SELECTION_ROUTER_NAME) {
-        this.popUpTo(UPDATE_ROUTINE_ROUTINE_ROUTER_NAME) {
+        this.popUpTo(UPDATE_ROUTINE_ROUTINE_SET_ROUTER_NAME) {
             inclusive = true
         }
     }
