+++ feature/update-routine/routine-selection/src/main/java/com/gradation/lift/feature/update_routine/routine_selection/UpdateRoutineRoutineSelectionNavigation.kt
@@ -6,7 +6,7 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.gradation.lift.navigation.Router
-import com.gradation.lift.navigation.navigation.navigateRoutineSelectionToRoutine
+import com.gradation.lift.navigation.navigation.navigateRoutineSelectionToRoutineSet
 import com.gradation.lift.navigation.navigation.navigateUpdateRoutineGraphToHomeGraph
 
 @RequiresApi(Build.VERSION_CODES.O)
@@ -16,8 +16,8 @@ fun updateRoutineRoutineSelectionScreen(
 ) {
     navGraphBuilder.composable(Router.UPDATE_ROUTINE_ROUTINE_SELECTION_ROUTER_NAME) {
 
-        val navigateRoutineSelectionToRoutine: () -> Unit =
-            { navController.navigateRoutineSelectionToRoutine() }
+        val navigateRoutineSelectionToRoutineSet: () -> Unit =
+            { navController.navigateRoutineSelectionToRoutineSet() }
 
 
         val navigateUpdateRoutineGraphToHomeGraph: () -> Unit =
@@ -25,7 +25,7 @@ fun updateRoutineRoutineSelectionScreen(
 
         UpdateRoutineRoutineSelectionRoute(
             navController,
-            navigateRoutineSelectionToRoutine,
+            navigateRoutineSelectionToRoutineSet,
             navigateUpdateRoutineGraphToHomeGraph
         )
     }
